+++ tests/MaxMind/Test/MinFraud/Model/EmailDomainTest.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace MaxMind\Test\MinFraud\Model;
+
+use MaxMind\MinFraud\Model\EmailDomain;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @coversNothing
+ */
+class EmailDomainTest extends TestCase
+{
+    public function testEmailDomain()
+    {
+        $array = [
+            'first_seen' => '2017-01-02',
+        ];
+        $email = new EmailDomain($array);
+
+        $this->assertSame(
+            $array['first_seen'],
+            $email->firstSeen,
+            'firstSeen'
+        );
+    }
+}